Tested with fresh system

+++ Server/middleware/Blockchain/Utilities/Snap.js
@@ -1,13 +1,11 @@
 const fs = require('fs')
 
 const Snapshot = () => {
-    var Chain = JSON.parse(fs.readFileSync('./middleware/Blockchain/Storage/Master.json'))
-    var now = new Date().getTime().toString()
-    fs.writeFileSync('./middleware/Blockchain/Storage/Snaps/' + now + '.json', JSON.stringify(Chain, null, 4))
-
-    return new Promise((resolve) => {
-        resolve()
-    });
+    if(fs.existsSync('./middleware/Blockchain/Storage/Master.json') && JSON.parse(fs.readFileSync('./middleware/Blockchain/Storage/Master.json')).Events.length > 0){
+        var Chain = JSON.parse(fs.readFileSync('./middleware/Blockchain/Storage/Master.json'))
+        var now = new Date().getTime().toString()
+        fs.writeFileSync('./middleware/Blockchain/Storage/Snaps/' + now + '.json', JSON.stringify(Chain, null, 4))
+    }
 }
 
 setInterval(Snapshot, 900 * 1000)
